Update startup_MK64F12.S
This file is a derived work from KSDK (eg. all the interrupt handler functions have the same names) but it does not have Freescale Copyright statement.  Surmise whoever in mbed did this work back in 2011 did not have the same interpretation of what counts as a "derived work" as we do now.  I have added back a Fresscale copyright statement.