Restore proper license on startup file
It looks like our original ARMCC startup file came from an older version of
K(P)SDK and didn't have any license. However, the file is very similar to the
corresponding file in KSDK 1.2.0, so this commit changes the license to
match the one in KSDK (and the license of the rest of files from KSDK
from this module).